split manager in its own file and more tests
