+++ pkg/websocket/connection.go
@@ -6,110 +6,16 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"net/http"
 	"sync"
 	"time"
 
 	ws "github.com/gorilla/websocket"
 	log "github.com/sirupsen/logrus"
+
+	"github.com/stripe/stripe-cli/pkg/stripeauth"
 )
 
-// ConnectionManager provides an interface for interacting with a *series*
-// of Connections; that establishes a new connection when the previous connection
-// disconnects.
-type ConnectionManager interface {
-	// Run accepts "onMessage" and "onFinish" handlers, so the caller can be
-	// notified of incoming messages and fatal errors. The caller must also pass a
-	// "connect" function that defines how to get a single Connection. Run returns
-	// a "send" function, which the caller should use to send outgoing messages out
-	// through the current connection.
-
-	Run(
-		ctx context.Context,
-		onMessage func([]byte),
-		onTerminate func(error),
-		connect func() (Connection, error),
-	) func(io.Reader)
-}
-
-type ConnectionManagerCfg struct {
-	requeueBackoffDuration time.Duration
-}
-
-func NewConnectionManager(cfg *ConnectionManagerCfg) ConnectionManager {
-	if cfg == nil {
-		cfg = &ConnectionManagerCfg{}
-	}
-	ret := &connectionManager{}
-	if cfg.requeueBackoffDuration == 0 {
-		ret.requeueBackoffDuration = time.Second
-	} else {
-		ret.requeueBackoffDuration = cfg.requeueBackoffDuration
-	}
-	return ret
-}
-
-type connectionManager struct {
-	requeueBackoffDuration time.Duration
-}
-
-func (c *connectionManager) backoff() {
-	time.Sleep(c.requeueBackoffDuration)
-}
-
-func (c *connectionManager) Run(
-	ctx context.Context,
-	onMessage func([]byte),
-	onTerminate func(error),
-	connect func() (Connection, error),
-) func(io.Reader) {
-	writes := make(chan io.Reader)
-	go func() {
-		for {
-			conn, err := connect()
-			if err != nil {
-				onTerminate(err)
-				return
-			}
-			onConnDisconnect := make(chan struct{})
-			sendToConn := conn.Run(ctx, onMessage, func() {
-				close(onConnDisconnect)
-			})
-			c.writeLoop(writes, sendToConn, onConnDisconnect)
-		}
-	}()
-	send := func(msg io.Reader) {
-		writes <- msg
-	}
-	return send
-}
-
-// writeLoop blocks, sending messages from the `writes` channel via the
-// `sendToConn` function, until onConnDisconnect is triggered.
-func (c *connectionManager) writeLoop(
-	writes chan io.Reader,
-	sendToConn func(io.Reader) error,
-	onConnDisconnect chan struct{},
-) {
-	for {
-		select {
-		case msg := <-writes:
-			err := sendToConn(msg)
-			if err != nil {
-				// Requeue after a backoff
-				go func() {
-					c.backoff()
-					writes <- msg
-				}()
-			}
-		case <-onConnDisconnect:
-			return
-		}
-	}
-
-}
-
-// Connection represents something such as a websocket connection, that can
-// transmit messages, receive messages, and die.
 type Connection interface {
 	// Run accepts "onMessage" and "onDisconnect" handlers, so the caller can be
 	// notified of incoming messages and disconnect events. It returns a "send"
@@ -208,10 +114,12 @@ func (c *connection) messageSender(onDisconnect func()) func(msg) error {
 		}
 
 		if m.isPing {
+			fmt.Println("INSIDE IS PING")
 			c.log.WithFields(log.Fields{
 				"prefix": "connection.sendMsg",
 			}).Debug("Sending ping message")
 			if err = c.conn.WriteMessage(ws.PingMessage, m.data); err != nil {
+				fmt.Println("ERROR SENDING PING")
 				if ws.IsUnexpectedCloseError(err, ws.CloseNormalClosure) {
 					c.log.Debug("write error: ", err)
 				}
@@ -248,14 +156,20 @@ func (c *connection) messageSender(onDisconnect func()) func(msg) error {
 }
 
 func (c *connection) pingPong(sendPing func() error, onDisconnect func()) {
+	fmt.Println("START PINGPONG")
 	stop := make(chan struct{})
 	ticker := time.NewTicker(time.Millisecond * 200)
-	defer func() {
-		ticker.Stop()
-	}()
+	fmt.Printf("tricker: %+v", ticker)
 
+	fmt.Println("GO FUNC")
 	go func() {
+		defer func() {
+			fmt.Println("DEFERED")
+			ticker.Stop()
+		}()
+
 		c.log.Debug("ticking...")
+		fmt.Println("FOR")
 		for {
 			select {
 			case <-stop:
@@ -263,6 +177,7 @@ func (c *connection) pingPong(sendPing func() error, onDisconnect func()) {
 				return
 			case <-ticker.C:
 				c.log.Debug("sending ping...")
+				fmt.Println("SENDING PING")
 				sendPing()
 				err := c.conn.SetWriteDeadline(time.Now().Add(c.writeWait))
 				if err != nil {
